Update auth.js
The statusText should not be hardcoded. It can be misleading if an error occurs in the try{} that is not related to jwtDecode(). For example, if a bad path is passed in for redirect. 

It is more helpful to the developer if the actual exception is printed out. In the event jwtDecode() fails, it will print "Error: Invalid token specified".